Added class Operation as interface. Moved OperationType from plugin to api. Added "addOperation" method from SelectionManager to SelectionController  Renamed Operation in plugin to OperationImpl and let it implement Operation  Modified SelectionManager to work with Operation interface
