+++ astraledit-bukkit-api/src/main/java/com/github/shynixn/astraledit/api/bukkit/business/controller/Operation.java
@@ -1,0 +1,24 @@
+package com.github.shynixn.astraledit.api.bukkit.business.controller;
+
+public interface Operation {
+    /**
+     * Returns the operationType
+     *
+     * @return type
+     */
+    OperationType getType();
+
+    /**
+     * Returns the unknown operation Data
+     *
+     * @return data
+     */
+    Object getOperationData();
+
+    /**
+     * Sets the data of the operation
+     *
+     * @param operationData operationsData
+     */
+    void setOperationData(Object operationData);
+}
