+++ astraledit-bukkit-api/src/main/java/com/github/shynixn/astraledit/api/bukkit/business/controller/OperationType.java
@@ -1,4 +1,4 @@
-package com.github.shynixn.astraledit.bukkit.logic.business;
+package com.github.shynixn.astraledit.api.bukkit.business.controller;
 
 /**
  * Copyright 2017 Shynixn
