+++ astraledit-bukkit-api/src/main/java/com/github/shynixn/astraledit/api/bukkit/business/controller/SelectionController.java
@@ -80,4 +80,6 @@ public interface SelectionController extends AutoCloseable {
      * @return controll
      */
     DependencyWorldEditService getWorldEditController();
+
+    void addOperation(Player player, Operation operation);
 }
