+++ astraledit-bukkit-plugin/src/main/java/com/github/shynixn/astraledit/bukkit/logic/business/command/MirrorCommand.java
@@ -3,8 +3,8 @@ package com.github.shynixn.astraledit.bukkit.logic.business.command;
 import com.github.shynixn.astraledit.api.bukkit.business.command.PlayerCommand;
 import com.github.shynixn.astraledit.bukkit.AstralEditPlugin;
 import com.github.shynixn.astraledit.bukkit.Permission;
-import com.github.shynixn.astraledit.bukkit.logic.business.Operation;
-import com.github.shynixn.astraledit.bukkit.logic.business.OperationType;
+import com.github.shynixn.astraledit.bukkit.logic.business.OperationImpl;
+import com.github.shynixn.astraledit.api.bukkit.business.controller.OperationType;
 import com.github.shynixn.astraledit.bukkit.logic.business.SelectionManager;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
@@ -38,7 +38,7 @@ public class MirrorCommand implements PlayerCommand {
                     player.sendMessage(AstralEditPlugin.PREFIX_ERROR + "You don't have a valid render.");
                 } else {
                     this.selectionManager.getSelection(player).mirror();
-                    this.selectionManager.addOperation(player, new Operation(OperationType.MIRROR));
+                    this.selectionManager.addOperation(player, new OperationImpl(OperationType.MIRROR));
                 }
             });
             return true;
