+++ astraledit-bukkit-plugin/src/main/java/com/github/shynixn/astraledit/bukkit/logic/business/command/PlaceCommand.java
@@ -4,8 +4,8 @@ import com.github.shynixn.astraledit.api.bukkit.business.command.PlayerCommand;
 import com.github.shynixn.astraledit.api.bukkit.business.entity.Selection;
 import com.github.shynixn.astraledit.bukkit.AstralEditPlugin;
 import com.github.shynixn.astraledit.bukkit.Permission;
-import com.github.shynixn.astraledit.bukkit.logic.business.Operation;
-import com.github.shynixn.astraledit.bukkit.logic.business.OperationType;
+import com.github.shynixn.astraledit.bukkit.logic.business.OperationImpl;
+import com.github.shynixn.astraledit.api.bukkit.business.controller.OperationType;
 import com.github.shynixn.astraledit.bukkit.logic.business.SelectionHolder;
 import com.github.shynixn.astraledit.bukkit.logic.business.SelectionManager;
 import org.bukkit.entity.Player;
@@ -45,7 +45,7 @@ public class PlaceCommand implements PlayerCommand {
                 player.sendMessage(AstralEditPlugin.PREFIX_ERROR + "You don't have a valid render.");
             } else {
                 player.sendMessage(AstralEditPlugin.PREFIX_SUCCESS + "Placing render ...");
-                final Operation operation = new Operation(OperationType.PLACE);
+                final OperationImpl operation = new OperationImpl(OperationType.PLACE);
                 operation.setOperationData(((SelectionHolder) selection).getTemporaryStorage());
                 selection.placeBlocks(() -> {
                     this.manager.addOperation(player, operation);
